Separate out Queue into its own module
