+++ azure-functions-sdk/src/commands/new.rs
@@ -17,9 +17,11 @@ use syn::{parse_file, Expr, ExprArray, Item};
 
 mod blob;
 mod http;
+mod queue;
 
 pub use self::blob::Blob;
 pub use self::http::Http;
+pub use self::queue::Queue;
 
 fn get_path_for_function(name: &str) -> Result<String, String> {
     if !Regex::new("^[a-zA-Z][a-zA-Z0-9_]*$")
@@ -265,81 +267,6 @@ impl<'a> From<&'a ArgMatches<'a>> for New<'a> {
             quiet: args.is_present("quiet"),
             color: args.value_of("color"),
             args,
-        }
-    }
-}
-
-
-
-struct Queue<'a> {
-    name: &'a str,
-    queue_name: &'a str,
-}
-
-impl<'a> Queue<'a> {
-    pub fn create_subcommand<'b>() -> App<'a, 'b> {
-        SubCommand::with_name("queue")
-            .about("Creates a new queue triggered Azure Function.")
-            .arg(
-                Arg::with_name("name")
-                    .long("name")
-                    .short("n")
-                    .value_name("NAME")
-                    .help("The name of the new Azure Function.")
-                    .required(true),
-            )
-            .arg(
-                Arg::with_name("queue_name")
-                    .long("queue-name")
-                    .short("q")
-                    .value_name("QUEUE")
-                    .help("The name of the storage queue to monitor.")
-                    .required(true),
-            )
-    }
-
-    pub fn execute(&self, quiet: bool) -> Result<(), String> {
-        Queue::validate_queue_name(self.queue_name)?;
-
-        let data = json!({
-            "name": self.name,
-            "queue_name": self.queue_name
-        });
-
-        create_function(self.name, "queue.rs", &data, quiet)
-    }
-
-    fn validate_queue_name(name: &str) -> Result<(), String> {
-        if name.len() < 3 {
-            return Err(format!(
-                "queue name '{}' must be at least 3 characters.",
-                name
-            ));
-        }
-
-        if name.len() > 63 {
-            return Err(format!(
-                "queue name '{}' cannot be more than 63 characters.",
-                name
-            ));
-        }
-
-        if !Regex::new("^[a-zA-Z0-9][a-zA-Z0-9-]*[a-zA-Z0-9]$")
-            .unwrap()
-            .is_match(name)
-        {
-            return Err(format!("queue name '{}' must start and end with a letter or number and can only contain letters, numbers, and hyphens.", name));
-        }
-
-        Ok(())
-    }
-}
-
-impl<'a> From<&'a ArgMatches<'a>> for Queue<'a> {
-    fn from(args: &'a ArgMatches<'a>) -> Self {
-        Queue {
-            name: args.value_of("name").unwrap(),
-            queue_name: args.value_of("queue_name").unwrap(),
         }
     }
 }
